feat: unify PercentSuffixTextField
